requirements.txt
numpy>=1.14.2
pandas>=0.22.0
h5py>=2.6.0
matplotlib>=2.2.2
netcdf4>=1.2.2
scipy>=1.0.0
requests>=2.9.1 
statsmodels>=0.8.0
pyramid-arima>=0.6.5
